Use fixed 32bit sign argument to error functions
Ideally sign should be bool, but we don't want to depend on stdbool.h
and sometimes (e.g. in powf) it's more efficient to pass a non-zero
value than 1 to indicate that the sign should be set.

While working on double precision versions i realized that the unsigned
long argument is fragile (may fail on 32bit targets only).