Add simple sincosf test support
The test code does not handle multiple return values so use wrappers
around sincosf.  The same behaviour is expected as with sinf and cosf
except the status flags should be the union of calling both sinf and
cosf for a particular input, for this reason separate name is used
for sincosf tests instead of reusing existing sinf and cosf ones.

The provided tests are copies of existing sinf and cosf tests and
a sincosf specific underflow test.