More portable default setting for HAVE_FAST_FMA
FP_FAST_FMA is the standard way to decide if fma is fast, unfortunately
in practice it can be defined even if fma is not inlined (e.g. with
-fno-builtin-fma), and it does not really help if the libc has a single
instruction implementation: the call overhead is too much.  Most of the
time it is the correct check though and without configure time checks
this is the closest we can get.