[ARMADA-2086] Fix performance regression in atan2f
A mistake in atanf_common.h meant we were using double-precision FMAs
where previously we were using single-precision. This patch changes
back to single-precision FMA, and adjusts the underflow threshold
accordingly, as well as catching some new underflows that occur as a
result. This brings performance close to where it was before, however
there is still a small penalty due to the new special-case, which was
missed before the original change took place.

Performance on N1:

Difference between the original single-precision implementation, which
did not handle underflows correctly (i.e. this performance difference
is almost completely due to the new special case):

   wide [      -10.0,   10.0]: 9.16 -->  9.99 (x1.09)
 narrow [       -1.0,    1.0]: 9.25 -->  9.98 (x1.08)
  pos_a [        0.0,    1.0]: 9.05 -->  9.95 (x1.10)
  pos_b [        1.0,  100.0]: 9.27 -->  9.98 (x1.08)
  pos_c [  1000000.0,  1e+32]: 9.15 -->  4.29 (x0.47)
  neg_a [       -0.0,   -1.0]: 9.11 -->  9.99 (x1.10)
  neg_b [       -1.0, -100.0]: 9.05 -->  10.0 (x1.10)
  neg_c [ -1000000.0, -1e+32]: 9.12 -->  4.29 (x0.47)
elefunt [        2.0,    2.0]:  9.1 -->  9.99 (x1.10)

Difference between the double-precision implementation and this patch:

   wide [      -10.0,   10.0]: 12.27 -->  9.99 (x0.81)
 narrow [       -1.0,    1.0]: 12.21 -->  9.98 (x0.82)
  pos_a [        0.0,    1.0]: 12.25 -->  9.95 (x0.81)
  pos_b [        1.0,  100.0]: 12.18 -->  9.98 (x0.82)
  pos_c [  1000000.0,  1e+32]: 12.27 -->  4.29 (x0.35)
  neg_a [       -0.0,   -1.0]: 12.27 -->  9.99 (x0.81)
  neg_b [       -1.0, -100.0]: 12.22 -->  10.0 (x0.82)
  neg_c [ -1000000.0, -1e+32]: 12.27 -->  4.29 (x0.35)
elefunt [        2.0,    2.0]: 12.24 -->  9.99 (x0.82)

Change-Id: Ic66da6f5a10df680b5a714a486e28ec4c89a93c6