Fix NaN behaviour in atan2
NaN handling in double-precision atan2 was not consistent with the
reference. This is now updated. Since atan2 is used as reference for
atan2f, we also have to update the NaN handling there. An explicit
test for correct -NaN handling has been added in runulp.sh

atan2 is used as a scalar fallback for vector variants of atan2 where
an input is NaN - the same problem was observed for those and is now
fixed.