pl/math: Add vector/SVE powi
Add double-precision SVE powi. powi is not required to meet any
particular accuracy requirements, so we do not test in the usual
way. Instead we check for bitwise reproducibility with a scalar
reference.

Vector powi does not follow the usual naming convention, instead the
trailing k denotes 64-bit integer, which distinguishes it from vector
powif.