string: arm: Add new functionality to prologue/epilogue assembler macros.
This patch adds options for automatic alignment enforcement and for
pushing/popping the lr register to prologue and epilogue assembler macros,
while making the pushing of the ip register optional for PACBTI.

Furthermore, as the use of these macros is independent of PACBTI and
may be used on architectures without the feature, the macros are moved
to a common header.

Improvements are also made to cfi handling. Where absolute cfi offset
calculation is complicated by optional function prologue
parameters (e.g. the pushing of pac-codes to the stack for M-profile
pacbti on function entry and pushing of dummy register when alignment
required), replace the use of .cfi_offset for .cfi_rel_offset,
simplifying cfi calculations by basing offsets on SP rather than the
cfa.

Finally, extensive in-source documentation is added to these macros to
facilitate their use and further development.

Built w/ arm-none-linux-gnueabihf, ran make check-string w/ qemu-arm-static.