pl/math: Add vector/Neon expm1f
New routine is a vector port of the scalar algorithm, with fallback to
the scalar variant for large and special input. This enables us to
simplify elements of the algorithm which were necessary for large
input. It also means that, as long as we fall back to the scalar for
tiny input as well (dependent on the value of WANT_ERRNO), the routine
sets fenv flags correctly. Some changes were needed in runulp.sh to
test this, as previously we have not been interested in fenv behaviour
of vector routines.