string: Compile memcpy-sve.S for aarch64 if compiler supports it
This is a partial revert of b7e368fb.
If SVE assembly is guarded by __ARM_FEATURE_SVE,
it cannot build when SVE is not enabled by the
build system. This is ok on AOR, but because
Android (bionic) uses ifuncs to select the
appropriate assembly at runtime, these
need to compile regardless of if the target
actually supports the instructions.

Check for AArch64 and GCC >= 8 or
Clang >= 5 so that SVE is not used
on compilers that do not support it.
This condition will always be true
on future builds of Android for
AArch64.