math: Avoid libc symbol name conflicts with static linking
For helper functions in the math code we want to use symbols in the
libc reserved namespace so the code can be upstreamed to a libc with
minimal changes.  However we don't want such symbols in libmathlib.a
when building Arm Optimized Routines since they may conflict with the
libc, so rename the symbols using macro redirects.