math: Cleanup Neon exp & expf
Replace implicit arithmetic operations with intrinsics and move data
to struct for improved memory access.

Also remove configure option to build double-precision exp with a
larger table. This enabled a significantly more accurate (1.04 ULP vs
2.38 ULP) routine, but was never used as this kind of accuracy is not
currently a requirement for vector routines. It has been removed for
the sake of simplicity.