Implementation of tgammal() suitable for 128-bit long double
This is a function that existing float128 maths libraries seem not to
have. For example, FreeBSD's libm on AArch64 implements tgammal by
simply casting to double and calling ordinary tgamma(), meaning that a
large range of input values (from about 171.63 to 1755.54) return
infinity when they could return a finite value, and the finite values
that are returned have only 52 bits of mantissa precision instead of
112.

This implementation is suitable for standalone use (it uses a
non-standard name and avoids non-portable dependency on endianness),
but could easily be integrated into a system libm.