pl/math: Rework Neon/f64 polynomial helpers
For small poly order, such as in exp10, better codegen may be obtained
by avoiding use of the helpers. For many routines a new array for
coeffs is needed, as they were previously sharing the scalar
array. Note this is NOT a cleanup patch, so no other constants have
been moved to data arrays, and the old way of ensuring good codegen,
namely marking the array is 'volatile' is used. All such routines are
due a cleanup, and this is an acceptable halfway point until then.