pl/math: Add 'zerofy' helper and use instead of BIC
'zerofy'-ing special lanes is a very common operation, especially for
WANT_SIMD_EXCEPT, where it is necessary to mask out special
lanes. This is generally done with BIC, which requires several
reinterprets, making operation difficult to understand (even though it
is so common). This change adds a helper to make the intent of such
masks clearer.