Fix erfinv table lookup indices
Fix table lookup indices in both single- and double-precision erfinv
implementations such that the base index for each vector element points
to the base of the byte table, rather than to the corresponding lane.
These indices are adjusted by conditionally applying an offset to read
the lane that we are interested in.