Add streaming compatible infrastructure and port easy cases
Second attempt at enabling a streaming-compatible (SC) port, using a
slightly cleaner approach which makes reusing existing test and bench
tools easier. Updated ulp and mathbench so that SC functions can be
tested with streaming mode enabled.

Added SC versions of all functions for which enabling streaming
compatibility is trivial (here trivial means that a routine can be
built and has no mode switch on the fast path, just by adding the SC
attribute to the function and its helpers in the same file). SC
versions with the same implementation are implemented using the same
approach we used to take for routines which shared implementation but
had different PCS - the SC source just toggles a flag on and includes
the non-compatible source.

Extended PL_SIG helpers, and also did some cleanup, to handle the new
routines.