test/ulp: Enable user to specify predicate for SVE routines
User can now pass a predicate to the ULP tool. The expected format is
as a uint64, where if bit N is 1 then the corresponding lane in the
predicate will be active, otherwise inactive (SVE max vector length is
64 lanes for single-precision). For example -p 2 tests only the second
lane, -p 0xf tests the first 4 lanes.

This revealed that SVE log2 and log2f were being tested for correct FP
exceptions, despite getting them wrong. Prior to this the flag had
been ignored, so this was not detectewd, however it now needs to be
removed.

Expose -p argument via 'pred' environment variable for runulp.sh, and
improve func handling.