pl/math: Improve codegen in SVE expf and inline helper
Improve special-case handling. Use inline helper to duplicate the
entire computation between the special- and non-special case
branches. For exp2f it is possible to avoid this by reconstructing x =
r + n, but there is a risk that the FEXPA and FACGT get emitted in the
wrong order and require MOVPRFX.

Remove further MOVPRFX and MOV by rearrange some MLAs and MLSs - by
making the multiplicand the destination we can avoid a MOVPRFX in
several cases. Also change which constants go in the vector used for
lanewise ops - the last lane is no longer wasted.

Also small tidying up of test intervals. Spotted that shift was
incorrect in exp2f and exp10f (w.r.t. to the comment that explains
it. Fixed - worst-case ULP for exp2f moves around but it doesn't
change significantly for either routine).