pl/math: Improve codegen in users of Neon log1pf helper
log1pf is quite register-intensive - use fewer registers for the
polynomial, and make various changes to shorten dependency chains in
parent routines. There is now no spilling with GCC 14. Accuracy moves
around a little - ULP limits and filenames adjusted accordingly.

Use the helper in log1pf as well, instead of having separate
implementations. The more accurate polynomial means special-casing can
be simplified, and the shorter dependency chain avoids the usual dance
around x0, which is otherwise difficult.

There is a small duplication of vectors containing 1.0f (or
0x3f800000) - GCC is not currently able to efficiently handle values
which fit in FMOV but not MOVI, and are reinterpreted to
integer. There may be potential for more optimisation if this is
fixed.