math, pl/math: Compile math for WoA.
Make sure the libary builds on Msys without error.
Tested with `make install`.
Building anything else fails (test, bench, ...), tests
are handled in the next patch.

Added OS detection in config (requires Unix compatibility).
The OS variable is only used in configuration.
In the source code we want to be more specific and use
- ifdef _M_ARM64 for WoA-specifics or
- ifdef _WIN32 for Windows specifics, e.g. if code is
  potentially cross-architecture, like scalar code.
- ifdef _MSC_VER for Visual Studio-specifics.
The reason is that sources are often used almost as is in
other projects that may not define the OS variable.