math, pl/math: Redefine INFINITY to work on Windows
Existing Windows definition of INFINITY might not produce the
result that is expected in math routines (in Z rounding mode
when compiled with -frounding-math).
On Windows with clang (or gcc - not yet available), we enforce
glibc math.h definition of INFINITY. MSVC does show this bug
therefore we leave INFINITY untouched.