math, pl/math: Move all self-contained vector routines to math/
'Self-contained' meaning all those which do not depend on a symbol
from another TU (except for scalar fallbacks). Poly helpers have to be
duplicated between PL and math, and some extra handling is needed in
test machinery for SVE and control values.

cbrt is not moved across because of unexplained ULP failures - this
will be dealt with separately.