math: Move max ULP and fenv spec to routine sources
Prior to merging the majority of vector routines from PL, we want to
align how the two directories handle test parameters. This is the
first stage of incorporating the machinery from PL for autogenerating
test scripts and avoiding conflicts in several long lists of routines.

Routines from math have an extra requirement compared to PL, which is
that scalar routines need testing in different rounding modes, which
may require different ULP limits - there is some extra work needed to
handle this.

Also rename some variables in pl/math/Dir.mk so that they do not cause
collisions with lists in math/Dir.mk.