math: Move test intervals to source files
Using the same preprocessor machinery as PL directory to run them in
runulp.sh. Scalar routines are tested in all rounding modes, and
vector routines in nearest only with sign-of-zero disabled.

No longer need to generate separate limits files for arch and generic
routines - one suffices, but instead we have to generate separate
lists of intervals. This means WANT_SIMD_TESTS no longer needs to be
passed to runulp as if ARCH is not aarch64 then these intervals are
not emitted (similar for WANT_EXP10_TESTS).

Update PL's runulp to be aligned as closely as possible, however due
to different directory structure there are a couple of differences
which are noted in comments.