math, pl/math: Move vector erf/erfc variants to math
Unfortunately it is not possible to avoid duplicating the tables
between the directories, as the scalar variants are not suitable for
mathlib.

However the special SVE tables could be removed, as by using a
combination of mask-and-add instead of the shift-based index
calculation the routines can share the same table as other variants
with no performance degradation.