math, pl/math: Move sinpi, cospi, sincospi to math
The scalar versions of these are not sufficiently accurate that they
can be linked ahead of GLIBC. They are still required for scalar
fallbacks though, so they are renamed with arm_math_ prefix. Trigpi in
general is only supported for aarch64 (this was previously implicit,
as PL only supports aarch64).

Wrappers for ulp and mathbench are also moved to math. Due to the
arm_math prefix they cannot use the TEST_* macros.