math: Raise minimum GCC version
Always build SVE routines by default - this means that on AArch64 we
always require GCC >= 10 or LLVM >= 5 for SVE ACLE. There is no explicit
check for compatible compiler, the SVE routines will just fail to build
if CC is too old.

There is a lot of clearing up that can be done, as there were previously
many typedefs and macros to work around the possibility that SVE, and
even AdvSIMD, might not be available.

As well, it was spotted that the signature of sincos, sincospi and modf
were wrong according to VFABI - fixed in the routines and wrappers.