math: Add vector annotations to mathlib.h
GCC may emit 'half-width' aliases for single-precision Neon routines,
so these are added to the GLIBC ABI config option. With mathlib.h
included, GCC may now vectorise to our vector math symbols with
-ffast-math.

AOR now has declarations for several scalar routines which it does not
actually provide (such as double-precision trig functions), however it
is has always been expected that mathlib is linked ahead of libm
(several vector routines are not usable otherwise), so this is
acceptable.

Declarations are only added for routines whose signature is currently
supported for autovec in GCC (anything with a pointer in the signature
is not considered).