math: Remove sve-math-cflags variable
Instead of adding the arch to the compile line, add it as a pragma for
all SVE routines, and also in the tools depending on WANT_SVE_TESTS
(to enable tests to still run on non-SVE machines). Note users will
still have to disable WANT_SVE_TESTS if on a machine that doesn't
support SVE (even if they are not on AArch64).

Because SVE is now enabled by including sv_math.h, it is important
that it is always included before any other header (especially headers
with inline helpers) so that inlining won't be blocked by mismatching
attributes.