Corrected comment in add sv_expf_inline.h
The sve f32 exp function is using a polynomial of
degree 5, not 6 as written in the comment. I modified
the comment accordingly. Moreover, the term y is not
used anywhere in the code, while the result is defined
as a function of poly(r), hence I changed it as well.