networking/test/chksum.c: fix undefined-function error.
include/networking.h defines __chksum_arm_simd conditionally, under
`#if (__arm__ && __ARM_NEON)`. But the test program refers to it under
just `#if __arm__`. So if you have __arm__ but not __ARM_NEON, you get
the reference, but not the definition to go with it, causing a link
failure.

Fixed by making the `#if`s in the test match the ones in the header.