math/aarch64/experimental: Fast inaccurate vector sinf and cosf
First attempt at faster sinf/cosf with relaxed accuracy.
The upper limit is generally set such that the top half
bits in the mantissa are correct, ie about 4096 ULP
for single precisions.

We obtain just under 3000 ULP by removing the third term in the
Payne-Hanek reduction (the second is crucial for accuracy) and
keeping only 2 terms in the polynomial.

For SVE, an implementation based on ftrig does not allow reducing
the polynomial as much while maintaining sufficient accuracy.
Some optimization done here could also be done to existing sinf/cosf.