fp: add a second version of single-precision multiply.
This version is intended for all the AArch32 targets other than the
very restricted v6-M style: Arm state, and Thumb state with a full set
of 32-bit instructions. The same assembly language can be made to
compile in both modes, with just the occasional #ifdef to work around
some particular Arm instruction not being available in Thumb.

As far as I know there's no short name in Arm architecture terminology
to describe that combination, so I've called the subdirectory `at32`,
for 'Arm or Thumb, with 32-bit instructions'.

There's no new test for this implementation, because it reuses
`test-fmul.c` introduced in the previous patch.