fp: single-precision comparisons.
This provides an implementation of every single-precision comparison
function defined in RTABI, although (as usual) with different names so
that they can be tested without clashing with an existing library's
implementations.

The RTABI comparison functions come in a lot of flavours: nine in
total. Six of them test for a specific condition (EQ, LT, LE, etc) and
return a boolean in r0; the other three return in the flags, with
varying semantics.

Most of these functions have a common skeleton, in the header file
fcmp.h. That skeleton function is parametrised by each actual .S file,
which controls which way round the operands go (one is reversed), what
is returned in r0 if anything, and what's returned in the event of
finding a NaN.

The exception is the boolean comparison testing 'unordered', which is
simple enough to just do directly without any help.