fp: conversions from single precision to integers.
This patch adds four separate functions, all taking a float as input,
and delivering signed and unsigned 32- and 64-bit integers.

These are the first functions added in this project which need to
worry about endianness: the two halves of a 64-bit integer are
returned in r0 and r1, but big- and little-endian Arm disagree about
which is which. To make this convenient I've introduced an endian.h
which defines register aliases ah,al for the high and low words of a
64-bit quantity. This will also be useful once we start on the double
precision operations, so endian.h also defines bh,bl for the registers
containing a _second_ 64-bit quantity, which will be used by double-
precision arithmetic functions taking two inputs.