f2lz, l2f: improved replacement for RSC in Thumb.
Instead of RSB to perform the reversed subtraction without carry
followed by SUBCC to conditionally decrement the result, it's better
to use MVN to invert the first operand and then ADC to add to that
with carry.

In Arm they'd take the same number of instructions, but the reason we
need to do this at all is that we're in Thumb, in which case the SUBCC
approach also has an invisible IT instruction. On Cortex-M3 that IT
can't be folded into the preceding RSB, because that has to be 32-bit
to avoid clobbering the carry flag. So switching to MVN+ADC reliably
saves a cycle.