fp: improve checking in test-fcmp.
This change adds two safety measures to test-fcmp which I realized the
need for while developing test-dcmp.

First: it's possible to accidentally return a combination of C and Z
flags that an ordinary comparison can't, namely C=0 but Z=1. The
comparison functions that return in the flags should never do this,
because it would confuse a caller assuming that HI, EQ and LO are
mutually exclusive. So the test should check for that flag
combination, and fail if it sees it.

Also, the asm wrappers that call flag-returning functions modify the
flags (obviously!), so I should have included "cc" in their clobber
lists, to tell the compiler that they do that. In test-dcmp the lack
of that flag caused wrong behavior at run time. In test-fcmp it hasn't
yet, but only by luck.