fp: double-precision division.
This operation comes with some auxiliary machinery for formal
verification of an error bound used in the implementation.

The basic idea of this ddiv implementation is to calculate an
_approximate_ 64-bit value of the quotient a/b, and to know what the
maximum possible approximation error is. Then we know a range within
which the true quotient must fall, of width about 1/16 ULP. In the
common case, that range doesn't cross a rounding boundary, so we can
round and return the result fast. Otherwise we must check more
carefully to see which side of the rounding boundary the quotient is
really on.

The bound on the approximation error is critical to this strategy.
Making it too small would cause a correctness bug; making it any
bigger than necessary would cost performance, by sending cases
unnecessarily to the slower code path. So the assembly language
implementation here is accompanied by a Python script that uses Gappa
and the Rocq Prover to formally verify the error bound claimed in the
code.

The Python script works by containing an equivalent implementation of
the same calculation done in the assembly language. This introduces a
risk of the assembly language not matching the Python. To gain
confidence that the two pieces of code really do match, ddiv.S can be
compiled with the -DDIAGNOSTICS option to insert printf calls in the
computation of the approximate reciprocal, and the Python script can
be run in an evaluation mode where a pair of input values is passed in
on the command line. Both of these modes print out the final
approximate reciprocal and also some important intermediate values.
This allows the two pieces of code to be cross-checked against each
other, to ensure all those values are bit-for-bit identical.