fp: double-precision comparisons.
These are very similar to the single-precision ones, with only a small
amount of extra complexity to handle the inputs spanning two
registers. The basic structure of fallback strategies is identical.

The biggest differences are that the check for NaNs is done by CMN
rather than TST (because the double-precision exponent mask won't fit
in an immediate field), and that in the `at32` versions I had to
introduce a slightly strange 'SetResultRegisterNE' macro, which the
single-precision version doesn't have at all.

(It's only used when you've found the upper words of the two inputs to
be equal and not tested the lower words yet. The Thumb1 version
doesn't need it either, because it replaces that conditional operation
with a branch round an ordinary `SetResultRegister`.)