fp: change register aliases ah,al,bh,bl to xh,xl,yh,yl.
These register aliases are used in all the fp functions that deal with
64-bit inputs and outputs (double precision floats and 64-bit ints),
to allow most of the code to be endianness-independent: ah/al refer to
r0/r1 but which way round depends on endianness, and similarly bh/bl
refer to r2/r3.

This worked fine in all my testing, because I tested with the clang
assembler. Unfortunately, with gas, a problem arises: `bl` is also an
instruction mnemonic! And while clang thinks that this line

```
bl .req r2
```

is a perfectly sensible `.req` directive making a register alias, gas
instead decides it's a `bl` instruction in which the address operand
has been replaced with nonsense, and gives a syntax error.

I didn't find any way to work around this by quoting or escaping the
`.req` directive, or finding an alternative directive spelling which
puts the alias name second. We _could_ work around it using `#define`
in place of `.req`, relying on the C preprocessor being case
sensitive (so that the register alias would always be spelled `bl` and
the instruction would have to be consistently `BL`), but that's a
recipe for a future maintainer getting very confused.

So instead I've just done a bulk rename of everything that talks about
the two operands to every FP function: they're now called `x` and `y`,
instead of `a` and `b`. So the aliases are now xh,xl and yh,yl, none
of which is an instruction. But I've also tried as best I can to
consistently rename the operands in all the code comments, even (for
consistency) in the single-precision code which doesn't use those
aliases. Of course the single letter `a` was also used many times as
the indefinite article, so I had to be careful about which subset of
them to replace with `x`.

In ddiv, this bulk rename had a knock-on effect (but also only on the
comments): the comments were already using the letter `x` for the
approximate reciprocal, so I had to rename that too, to `r`.

None of this should cause any functional change (in any context where
the previous code would assemble successfully at all). All that's
changed is comments and identifier names.