fp: further build fixes for the GNU assembler.
When gas is assembling in Arm state, it won't accept the `.N` or `.W`
suffixes on instructions. (By contrast, clang just ignores them.)

The `.W` suffixes were all unnecessary anyway, so I've just removed
them. The one use of `.N`, in i2f.s, was deliberately arranged to save
a cycle on Cortex-M3, so I've left it in under #ifdef.

Test-building using the platform toolchain on AArch32 Debian bullseye
(gcc 10.2.1, binutils 2.35.2), I found that you have to take a bit of
care to ensure that the `#ifdef __thumb__` statements agree with the
mode the assembler is actually using. In particular, on the gcc driver
command line, `-mthumb` seems to control whether the preprocessor
defines `__thumb__`, whereas `-Wa,-mthumb` controls whether the
assembler is in Thumb state, so you need to specify both or neither.
However, after doing that, I was able to build all of these functions
successfully in both states.