fp: add GNU-style comparison functions.
The previously provided entry points correspond to the comparison
functions defined in RTABI. These all either return a boolean or
return in the flags. But the GNU runtime library specification defines
some comparison functions returning a 3-way result distinguishing x<y,
x=y and x>y. In this commit I add some more comparison variants
suitable for that use case.

GNU defines a three-way single-precision comparison function __cmpsf2,
which returns an integer <0, ==0 or >0 (similar to strcmp and memcmp).
I've added this under the name arm_fp_fcmp_gnu_cmp, and the
corresponding __cmpdf2 for double precision as arm_fp_dcmp_gnu_cmp.

The GNU runtime specification also defines six functions each
corresponding to one of the C comparison operators. Officially the
semantics of these are that you use the same comparison operator to
compare its return value with 0, and nothing else about the output is
guaranteed:

x == y   is implemented as   __eqsf2(x,y) == 0
x != y   is implemented as   __nesf2(x,y) != 0
x >  y   is implemented as   __gtsf2(x,y) >  0
x >= y   is implemented as   __gesf2(x,y) >= 0
x <  y   is implemented as   __ltsf2(x,y) <  0
x <= y   is implemented as   __lesf2(x,y) <= 0

However, the usual implementation approach is to make most of these
aliases to __cmpsf2. Since __cmpsf2 is defined to return >0 for
unordered results, effectively lumping 'unordered' in with 'greater',
it's correct to make __eqsf2, __nesf2, __ltsf2 and __lesf2 aliases to
it. However, doing this with __gesf2 and __gtsf2 would give the wrong
answer for NaNs. So I've provided a variant arm_fp_fcmp_gnu_rcmp,
which works exactly the same but reverses the NaN handling, so that
__gesf2 and __gtsf2 can be implemented as aliases to that.

This covers all the GNU comparison functions except for __unordsf2,
which is the same as fcmp_bool_un which is in this collection already.

In principle, it would also be possible to provide a full set of
dedicated functions implementing the semantics of each of the six
comparisons above, saving one or two instructions by only
distinguishing two possible return values rather than three. I haven't
done that in this commit, because it seems to me that the speed gain
is extremely marginal (typically only one instruction), and the space
cost of having 12 separate copies of the main comparison code is
likely to be more important. Also, an implementation hyper-optimized
in that way would be extra likely to trigger bugs in code generation
which had depended on non-guaranteed parts of the semantics.