math/aarch64/advsimd: Optimize pow(r)f.
Finally account for simplifications allowed by powr's definition.

Provide helpers with and without sign_bias so that
- powf can use sign bias in low path and no sign bias in fast path
- powrf never uses sign bias

Moved some helpers into top level source and only keep core computation
in inline helper files.

Question mark on supporting negative x outside of fast path for powrf.
It is probably not that easy to do it efficiently and does not cost
much to leave it in slow path.